feat(partial): Break out visualizations into their own pages w/ an index page.  In progress.
